+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Attribute, Category, Family and Option normalizers are now in the DI
  - Category export only create/update categories from the exported channel
  - Make it compatible with EE-1.11 and EE-1.12
+ - Add a log to profile calls to Magento
 
 ## BC Breaks
  - All DeltaExportBundle dependencies should be replaced by MagentoConnectorBundle ones